chore: extract different views in authenticate bank step to own components
