+++ src/client/v2/src/components/AddAccountModal/steps/AuthenticateBankStep/AuthenticateBankStep.module.css
@@ -2,60 +2,3 @@
   height: 100%;
 }
 
-.loaderSection {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: var(--mantine-spacing-sm);
-  text-align: center;
-}
-
-.instructionsSection {
-  width: 100%;
-  padding: var(--mantine-spacing-md) var(--mantine-spacing-xl);
-  background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-6));
-  border: 1px solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
-  border-radius: var(--mantine-radius-md);
-}
-
-.instructionItem {
-  display: flex;
-  align-items: flex-start;
-  gap: var(--mantine-spacing-sm);
-}
-
-.successMessage {
-  display: flex;
-  align-items: center;
-  gap: var(--mantine-spacing-sm);
-}
-
-.successSection {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: var(--mantine-spacing-md);
-  text-align: center;
-}
-
-.failureSection {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: var(--mantine-spacing-md);
-  text-align: center;
-}
-
-.bankHeader {
-  width: 100%;
-  padding: var(--mantine-spacing-sm) var(--mantine-spacing-md);
-  background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-6));
-  border: 1px solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
-  border-radius: var(--mantine-radius-md);
-}
-
-.accountCard {
-  padding: var(--mantine-spacing-md);
-  border: 1px solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
-  border-radius: var(--mantine-radius-md);
-}
