+++ src/client/v2/src/components/AddAccountModal/steps/AuthenticateBankStep/sub-components/BankAccountOverview/BankAccountOverview.module.css
@@ -1,0 +1,13 @@
+.bankHeader {
+  width: 100%;
+  padding: var(--mantine-spacing-sm) var(--mantine-spacing-md);
+  background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-6));
+  border: 1px solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
+  border-radius: var(--mantine-radius-md);
+}
+
+.accountCard {
+  padding: var(--mantine-spacing-md);
+  border: 1px solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
+  border-radius: var(--mantine-radius-md);
+}
