+++ src/client/v2/src/components/AddAccountModal/steps/AuthenticateBankStep/sub-components/BankAuthenticationFailed/BankAuthenticationFailed.module.css
@@ -1,0 +1,7 @@
+.failureSection {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: var(--mantine-spacing-md);
+  text-align: center;
+}
